BigDecimal refactor

+++ src/main/java/rhsouthsystem/desafioreceita/service/ReceitaService.java
@@ -3,10 +3,12 @@ package rhsouthsystem.desafioreceita.service;
 import org.springframework.stereotype.Service;
 import rhsouthsystem.desafioreceita.model.Conta;
 
+import java.math.BigDecimal;
+
 @Service
 public class ReceitaService {
 
-    public Conta atualizarConta(String agencia, String numeroConta, double saldo, String status) {
+    public Conta atualizarConta(String agencia, String numeroConta, BigDecimal saldo, String status) {
         Conta conta = new Conta();
 
         conta.setAgencia(agencia);
